fix: fix uid to int64 from int

+++ internal/repository/repository.go
@@ -31,7 +31,7 @@ type UsersTasks interface {
 	Create(user *models.UsersTasks) (int, error)
 	Get(id int) (*models.UsersTasks, error)
 	GetAll() ([]models.UsersTasks, error)
-	GetWhereUid(userId int) ([]models.UsersTasks, error)
+	GetWhereUid(userId int64) ([]models.UsersTasks, error)
 	Update(id int, usersTasks *models.UsersTasks) error
 	Delete(id int) error
 }
